Add generateQuiz() function. Solve selectQuestions() particular case.
Create generateQuiz(), the function that will call all of our
filters (year, difficulty etc.) to generate a balanced quiz
according to user input specifications.

Also, selectQuestions() had a problem when the input pool of
questions was smaller than the required size of the output.
Solve this situation by returning the whole input in this scenario.